feat: - Implemented SGD and Adam optimizers in `optimizer.c` with state management for tensors. - Created training loop functions in `trainLoop.c` for forward and backward passes, supporting both SGD and Adam optimizers. - Compared accuracy and speed between C model and PyTorch model

+++ tests/testUtils.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
-#include "../include/testUtils.h"
+#include "tests.h"
 
 void randomDataFill(Tensor *tensor) {
     for (int i = 0; i < tensor->rows * tensor->cols; i++)
-        tensor->data[i] = ((float)rand() / RAND_MAX - 0.5f) * 0.2f;
+        tensor->data[i] = ((float)rand() / (float)RAND_MAX - 0.5f) * 0.2f;
 }
